test: add tests for SntpClient
these tests compare the output of the Kotlin implementation
used in production with the output of the old Java implementation,
(which is now only kept in the test source set).

The Java version is believed to be correct, so if the outputs don't match,
there's a bug with the Kotlin implementation.